Limits the area on the map the user can pan
This prevents the user from panning the map to far in any direction. It
is a workaround to the fact that OpenLayers 3 does not support wrapping
the world when using a Vector layer.

This commit works by setting the "extent" option on the View. This is an
experimental feature of OpenLayers, but it seems to work well enough.
The feature works by setting the upper and lower bounds on the center of
the map. I had to experiment to find a set of values that would help
keep most of the map visible while also allowing the user to zoom in on
locations near the edges of the map (such as Hawaii and New Zealand).