Adds script to fetch event details
The script is written in Groovy and scrapes the coderetreat.org site for
basic event details. It then uses various Google Maps APIs to geocode
the locations and to identify the time zones of each location.

The script produces a CSV file suitable for use in a spreadsheet, and a
json file that the website can use to populate the GDCR map.

In order to use the script, you must have a Google API Key. See
https://github.com/googlemaps/google-maps-services-java#api-keys for
instructions on how to setup an API key.

Note: Since the script uses both the Geocoding and the Time Zone APIs,
you must enable both APIs in the Google Developer Console.